Extract manual syncing logic out of wallet RPC
